External file reading and parsing is done!
Note that this is very dirty. I think the input document has about three to four copies of itself in memory. One to read it in, one with modified contents, one modified into XML elements, and one that might be made when associating the element with the page document. Really, we should be reading in one line at a time, parsing that line, and injecting it directly into the page object.