Another messy commit
Designing and coding when you're tired is a dumb idea. I started working on
adding a title in this thing, but I haven't plumbed the needed data to the
theme to allow that to work.

Giant refactor needed so that the base wiki file has some meta-data like
title, author, and everything that's expected in the XML output. At this
point, it's likely that the only difference between the real output and
the base wiki file is that the contents are going to be stored in wiki
markup as opposed to HTML elements